refactor dpm campaign lifecycle routes
